+++ packages/uniforms-unstyled/src/SubmitField.js
@@ -2,9 +2,9 @@ import BaseField      from 'uniforms/BaseField';
 import React          from 'react';
 import filterDOMProps from 'uniforms/filterDOMProps';
 
-const SubmitField = ({inputRef, value, ...props}, {uniforms: {error, state: {disabled}}}) =>
+const SubmitField = ({disabled, inputRef, value, ...props}, {uniforms: {error, state}}) =>
     <input
-        disabled={!!(error || disabled)}
+        disabled={disabled === undefined ? !!(error || state.disabled) : disabled}
         ref={inputRef}
         type="submit"
         value={value}