+++ packages/uniforms/src/QuickForm.js
@@ -1,5 +1,5 @@
-import React       from 'react';
-import {PropTypes} from 'react';
+import PropTypes from 'prop-types';
+import React     from 'react';
 
 import BaseForm from './BaseForm';
 import nothing  from './nothing';