+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## Next
+## [v1.18.2](https://github.com/vazco/uniforms/tree/v1.18.1) (2017-06-09)
+- **Fixed:** `SubmitField` respects `disabled` prop.
+
+## [v1.18.1](https://github.com/vazco/uniforms/tree/v1.18.1) (2017-05-28)
 - **Changed:** Optimized `BaseField` props calculation.
+- **Changed:** Switched to `prop-types`.
 - **Fixed:** `Bridge.check` error message.
 
 ## [v1.18.0](https://github.com/vazco/uniforms/tree/v1.18.0) (2017-04-30)